+++ packages/reaxes-vue2/f.d.ts
@@ -1,13 +1,13 @@
 /*global provider*/
-declare const _ : typeof import('lodash');
-declare const React : typeof import('react');
-declare const useState : typeof React.useState;
-declare const useEffect : typeof React.useEffect;
-declare const useRef : typeof React.useRef;
-declare const useLayoutEffect : typeof React.useLayoutEffect;
-declare const useMemo : typeof React.useMemo;
-declare const useCallback : typeof React.useCallback;
-declare const utils: typeof import('reaxes-utils');
+// declare const _ : typeof import('lodash');
+// declare const React : typeof import('react');
+// declare const useState : typeof React.useState;
+// declare const useEffect : typeof React.useEffect;
+// declare const useRef : typeof React.useRef;
+// declare const useLayoutEffect : typeof React.useLayoutEffect;
+// declare const useMemo : typeof React.useMemo;
+// declare const useCallback : typeof React.useCallback;
+// declare const utils: typeof import('reaxes-utils');
 
 /*DOM*/