+++ packages/reaxes/f.d.ts
@@ -1,19 +1,19 @@
 /*global provider*/
-declare const _ : typeof import('lodash');
-declare const React : typeof import('react');
-declare const {
-	useEffect ,
-	useState ,
-	useLayoutEffect ,
-	Component ,
-	useRef ,
-	useMemo ,
-	useContext ,
-	useReducer ,
-	useCallback,
-}: typeof React;
-declare const utils : typeof import("reaxes-utils");
-declare const {orzPromise,crayon} : typeof utils;
+// declare const _ : typeof import('lodash');
+// declare const React : typeof import('react');
+// declare const {
+// 	useEffect ,
+// 	useState ,
+// 	useLayoutEffect ,
+// 	Component ,
+// 	useRef ,
+// 	useMemo ,
+// 	useContext ,
+// 	useReducer ,
+// 	useCallback,
+// }: typeof React;
+// declare const utils : typeof import("reaxes-utils");
+// declare const {orzPromise,crayon} : typeof utils;
 
 /*DOM*/